scripts/requirements.txt
pyyaml==6.0.2
praw==7.8.1